[VIEW] Documentation, CMakeLists changed to multiple, and here and there some changes

+++ src/logic/Stopwatch.hpp
@@ -14,7 +14,7 @@ public:
     void Start();
     void Tick();
 
-    [[nodiscard]] double GetFrameTime();
+    [[nodiscard]] float GetFrameTime();
 
 private:
     Stopwatch() = default;
